GpsListCell rename
