+++ databrowser/src/domain/cellComponents/components/cells/gpsPointsCell/GpsPointsCell.vue
